+++ src/app/pages/project/milestone/list-milestone/list-milestone.component.ts
@@ -33,11 +33,11 @@ export class ListMilestoneComponent implements OnInit {
   projectId: number;
   suites: TestSuite[];
 
-  public defSort = { property: 'name', order: TFOrder.desc };
+  public defSort = { property: 'due_date', order: TFOrder.asc };
 
   async ngOnInit() {
     this.projectId = this.route.snapshot.params.projectId;
-    this.suites = await this.suiteService.getTestSuite({project_id: this.projectId});
+    this.suites = await this.suiteService.getTestSuite({ project_id: this.projectId });
     this.milestones = await this.getMilestones();
     this.canEdit = await this.permissions.hasProjectPermissions(this.projectId,
       [EGlobalPermissions.manager], [ELocalPermissions.manager, ELocalPermissions.engineer]);
@@ -49,12 +49,20 @@ export class ListMilestoneComponent implements OnInit {
         type: TFColumnType.text,
         class: 'fit',
       }, {
+        name: 'Active',
+        property: 'active',
+        filter: true,
+        type: TFColumnType.checkbox,
+        editable: this.canEdit,
+        notEditableByProperty: { property: 'active', value: false }
+      }, {
         name: 'Name',
         property: 'name',
         filter: true,
         sorting: true,
         type: TFColumnType.text,
         editable: this.canEdit,
+        notEditableByProperty: { property: 'active', value: false },
         creation: {
           creationLength: 500,
           required: true
@@ -62,13 +70,23 @@ export class ListMilestoneComponent implements OnInit {
       }, {
         name: 'Suites',
         property: 'suites',
+        filter: true,
         type: TFColumnType.multiselect,
         editable: this.canEdit,
+        notEditableByProperty: { property: 'active', value: false },
         lookup: {
           entity: 'suites',
           propToShow: ['name'],
           values: this.suites,
         }
+      }, {
+        name: 'Due Date',
+        property: 'due_date',
+        sorting: true,
+        type: TFColumnType.date,
+        format: 'MMM dd, yyyy',
+        notEditableByProperty: { property: 'active', value: false },
+        editable: this.canEdit
       }];
   }
 
@@ -114,6 +132,8 @@ export class ListMilestoneComponent implements OnInit {
       this.milestones = await this.getMilestones();
     }
   }
+
+
 
   wasClosed(state: boolean) {
     this.hideModal = state;