+++ src/app/shared/models/milestone.ts
@@ -5,4 +5,6 @@ export class Milestone {
   name?: string;
   project_id?: number;
   suites?: TestSuite[];
+  active?: boolean | number;
+  due_date?: Date;
 }